fix: store and show package durations as raw hours
Package durations (duration.minHours/maxHours, durationOfSetup,
durationPerPerson) are stored as raw hours everywhere, decimals allowed
(1h30m = 1.5). Customer pages no longer treat them as minutes; they now
format as "1 hr 30 min" / "12 hrs" via a shared formatHours helper,
which the vendor publish summaries also use. Vendor flows save with
parseFloat so fractional hours are not truncated on re-save. Removes the
now-unused formatMinutes helper.

+++ src/features/packages/flows/venue/VenueFlow.tsx
@@ -395,8 +395,8 @@ export default function VenueFlow({ onExitFlow }: { onExitFlow?: () => void }) {
                     eventCategories: eventCategories ? eventCategories.split(',').map(s => s.trim()) : ['Wedding'],
                     poc: poc,
                     duration: {
-                        minHours: parseInt(minDuration) || 0,
-                        maxHours: parseInt(maxDuration) || 0,
+                        minHours: parseFloat(minDuration) || 0,
+                        maxHours: parseFloat(maxDuration) || 0,
                     },
                     crewSize: {
                         minPeople: parseInt(crewSize) || 0,
